octets: move tests to integration files
Move the public API tests out of src/lib.rs into octets/tests so
the crate code no longer carries a large inline test module. Keep the
Huffman-specific coverage in its own feature-gated test file.
